requirements.txt
Django==1.8.2
django-multiselectfield==0.1.3
django-selectable==0.9.0
django-smart-selects==1.1.1
django-wpadmin==1.7.4
Pillow==2.9.0
psycopg2==2.6.1
sorl-thumbnail==12.2
